Abstract shared logic in Crash and Location views into DataTable component
