+++ atd-vze/src/views/Locations/locationDataMap.js
@@ -4,25 +4,29 @@ const locationDataMap = [
     fields: {
       location_id: {
         label: "Location ID",
+        editable: false,
       },
       description: {
         label: "Description",
+        editable: false,
       },
       last_update: {
         label: "Last Update",
+        editable: false,
         format: "datetime",
       },
-      est_comp_cost: {
-        label: "Estimated Comprehensive Cost",
-        format: "dollars",
-      },
-      asmp_street_level: {
-        label: "ASMP Street Level",
-        editable: true,
-        uiType: "select",
-        lookupOptions: "atd_txdot__asmp_level_lkp",
-        lookupPrefix: "asmp_level",
-      },
+      // est_comp_cost: {
+      //   label: "Estimated Comprehensive Cost",
+      //   editable: false,
+      //   format: "dollars",
+      // },
+      // asmp_street_level: {
+      //   label: "ASMP Street Level",
+      //   editable: true,
+      //   uiType: "select",
+      //   lookupOptions: "atd_txdot__asmp_level_lkp",
+      //   lookupPrefix: "asmp_level",
+      // },
     },
   },
 ];
